+++ stxt/src/extension/AnalysisResult.ts
@@ -4,4 +4,5 @@ import type { StxtToken } from './Tokens';
 export interface AnalysisResult {
 	tokens: StxtToken[];
 	nodeByLine: Map<number, Node>;
+	commentLines: Set<number>;
 }